feat: use right data type for coords, log to a logs folder

+++ litter-tracker.Objects/StoreObjects/LitterPin.cs
@@ -12,7 +12,7 @@ namespace litter_tracker.Objects.StoreObjects
 
     public class LatLng
     {
-        public long Latitude { get; set; }
-        public long Longitude { get; set; }
+        public decimal Latitude { get; set; }
+        public decimal Longitude { get; set; }
     }
 }
